02-use-cases/05-Healthcare-Appointment-Agent/requirements.txt
PyYAML==6.0.2
requests==2.32.4
requests-auth-aws-sigv4==0.7
python-dotenv==1.1.0
mcp[cli]==1.9.4
strands-agents==0.1.8
ipykernel==6.29.5
aiohttp==3.12.13
langchain==0.3.26
langchain-aws==0.2.27
langchain-mcp-adapters==0.1.7
langgraph==0.5.1
./local_wheels/botocore-1.39.5-py3-none-any.whl
./local_wheels/boto3-1.39.5-py3-none-any.whl
./local_wheels/amzn_AWSAgentCredentialProviderServiceBotocoreClient-1.0-py3-none-any.whl
